--- src/main/java/kr/where/backend/group/dto/GroupCreateRequestDTO.java
@@ -1,16 +1,0 @@
-package kr.where.backend.group.dto;
-
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class GroupCreateRequestDTO {
-    @NotNull
-    private Long memberId;
-
-    @NotEmpty
-    private String groupName;
-}
