+++ src/main/java/kr/where/backend/group/dto/group/UpdateGroupDto.java
@@ -1,18 +1,15 @@
-package kr.where.backend.group.dto;
+package kr.where.backend.group.dto.group;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
-public class CreateGroupDto {
+public class UpdateGroupDto {
     @NotNull
-    private Long memberIntraId;
+    private Long groupId;
     @NotEmpty
-    @Size(max = 40)
     private String groupName;
 }
