+++ src/main/java/kr/where/backend/group/dto/groupmember/GroupMemberResponseDTO.java
@@ -1,4 +1,4 @@
-package kr.where.backend.group.dto;
+package kr.where.backend.group.dto.groupmember;
 
 import java.util.List;
 import lombok.Getter;
