+++ src/main/java/kr/where/backend/group/dto/groupmember/RequestGroupMemberDTO.java
@@ -1,4 +1,4 @@
-package kr.where.backend.group.dto;
+package kr.where.backend.group.dto.groupmember;
 
 import lombok.Builder;
 import lombok.Data;
