+++ src/main/java/kr/where/backend/group/dto/groupmember/ResponseGroupMemberDTO.java
@@ -1,4 +1,4 @@
-package kr.where.backend.group.dto;
+package kr.where.backend.group.dto.groupmember;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
